fix: typecheck and lint fixes
- mypy: implicit Optional, no-any-return, attr-defined fixes
- ruff: remove unused imports, raise from err, isort
- 19 mypy errors resolved, 11 ruff errors resolved

+++ src/infrastructure/entrypoints/cli_adapter.py
@@ -30,9 +30,9 @@ class CLIAdapter:
     def _show_progress_bar(self, current: int, total: int, width: int = 40) -> None:
         """Muestra barra de progreso."""
         if total == 0:
-            percent = 0
+            percent = 0.0
         else:
-            percent = current / total
+            percent = float(current / total)
 
         filled = int(width * percent)
         bar = '█' * filled + '░' * (width - filled)
